network_WiFi_{Hidden,Visible}Scan: more precise probe request capture
For a visible network, there are various reasons a station might not
need to send a probe request before authenticating with an AP (e.g.,
passive scanning). So network_WiFi_VisibleScan shouldn't be requiring a
non-zero number of probe requests in the packet capture.

Also, we should start our packet captures before we start configuring
the AP here, if we want to see all activity. For one, we could miss an
in-progress scan done by the station (which may find the AP) if we wait
too long to start the capture.

This should help decrease the number of (likely false) failures with
messages like:

  Expected exactly two SSIDs, but got frozenset([])

And we remove this failure case from network_WiFi_VisibleScan entirely:

  Expected SSIDs (['']), but got frozenset([])

BUG=chromium:627552
TEST=$subject tests, on lab devices with both whirlwind and stumpy APs

Change-Id: Ie4a1647d902402b1ff9fb686e705051f3947bc30
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/861200
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>