autotest: LocalSuite subclass for VM test retries
New subclass of Suite with methods mimicking but not duplicating the existing
schedule and result-retry flow, to handle retries for a local suite.

handle_local_result and _retry_local_result take an AFE job ID rather than a
Result object, and do not take a Waiter instance.

In RetryHandler, _should_retry_local_job also takes an ID rather than a Result,
and checks fewer considerations; it takes for granted that the job was not
aborted, and does not filter by _retry_level.

In other respects they closely mirror the behavior of _handle_result,
_retry_result, and _should_retry.

BUG=chromium:750918
TEST=suite:dummyflake retried, unit tests, and tryjob.

Change-Id: I68bf950b389bc247a9c1c1d6f9f7711023a13c10
Reviewed-on: https://chromium-review.googlesource.com/860163
Commit-Ready: Jacob Kopczynski <jkop@chromium.org>
Tested-by: Jacob Kopczynski <jkop@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>