firmware_Cr50ConsoleCommands: account for open drain
DIOB7 is open drain. Certain boards dont have this signal stuffed.
Because it's open drain, cr50 may be driving it high, but it will remain
low. Change the output to accept 0 or 1 on the read. This is not really
under cr50's control. We just care that it attempts to drive it high.

BUG=none
BRANCH=none
TEST=none

Change-Id: I03f8548639e2cdc8092c9612510bf6dcfb629118
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/862624
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>