Add detectors used in autotest-capability
autotest-capability is used to manage the autotest labels applied to the DUTs.
These labels define the capability of the DUTs so that each test can decide
if the test should run.
The detectors dynamically detects a capability on DUT in autotest-capability,
for example, kepler and CPU type.

BUG=chromium:789866
TEST=emerge daisy_spring, kevin, cyan, rikku and eve

Change-Id: I1cd3bb12d434959b8e5a61cf9db6f47279e020b0
Reviewed-on: https://chromium-review.googlesource.com/869150
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Kuang-che Wu <kcwu@chromium.org>
Reviewed-by: Owen Lin <owenlin@chromium.org>