network_WiFi_BSSTMReq: Add BSS Transition Management Request autotest
802.11v allows APs to send BSS Transition Management Requests to clients
asking them to roam to a different AP. This can be for a variety of reasons
including load balancing, imporving traffic quality, etc. This change adds an
autotest that brings up two APs, has the client connect to one of them, and
has that AP send a transition management request to the client, indicating
the second AP as a transition candidate. The expected behavior is that the
client successfully roams to the second AP.

BUG=chromium:836381
TEST=Manually tested
CQ-DEPEND=CL:876785
Change-Id: If33f9085d3a13e17316f0ebb547af9cdb2ee9eff
Reviewed-on: https://chromium-review.googlesource.com/876530
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>