[autotest] Double fork lucifer processes to not leak zombies
monitor_db doesnt reap its children, so we end up creating zombies by
forking once.  Double fork so we reparent to init.

Were not hugely concerned with lucifer trying to regain a tty, so
theres no need to setsid in the intermediate process, and we want the
new process to be the session/group leader for simplicity if we ever
need to SIGTERM.

BUG=chromium:803682
TEST=None

Change-Id: Id0317f56b401f7c53ea5b49cae3d73333557a9ee
Reviewed-on: https://chromium-review.googlesource.com/879394
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>