[autotest] Adjust DUT version labels verification.
Compare lsbrelease's CHROMEOS_RELEASE_BUILD_PATH againt the version
label to see if the correct version is installed, instead of
CHROMEOS_RELEASE_VERSION. A simple string compare makes the lab
slightly more robust against changes to the format of that string.

This does NOT affect the provision flow and verifications done there.

BUG=chromium:800562
TEST=None

Change-Id: Ia98c62744d1cbbacc22cb10500ae0e31b6e61755
Reviewed-on: https://chromium-review.googlesource.com/907326
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>