firmware_Cr50CCDServoCap: change acceptable I2C states
The I2C gpio is open drain. We don't have a lot of control on whether it
has the necessary pull up to make Cr50 think it is deasserted. Accept
the state that Cr50 thinks I2C is enabled even though it probably isn't.
If Cr50 correctly detects servo and disables the uart TX signals and the
SPI signal it's probably also disabling I2C.

BUG=none
BRANCH=none
TEST=run on Dru

Change-Id: I7619c2ac1ca515f0e922424d0acba6e971e80a52
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/882381