Add OOBE delta test and skip rialto failure.
The test is failing on one device (veyron_rialto) because it has its own
OOBE that it automatically starts stepping through. We will skip the
test on this device.

Add a test to force OOBE with a delta payload.

BUG=None
TEST=autoupdate_ForcedOOBEUpdate and autoupdate_ForcedOOBEUpdate.delta
pass. Test is skipped on rialto.

Change-Id: I516660cfda28bd901043cbf1e594eb06c59936e2
Reviewed-on: https://chromium-review.googlesource.com/882213
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>