power_status: Fix jitter in MeasurementLogger
Currently when we ask MeasurementLogger to log data every
20 seconds, Measurement logger will do the following

while not done
  log power
  sleep(20 seconds)

This lead to timing jitter as logging power does not finish
instantly and logging timestamp will be something like this.
0, 20.001, 40.003, 60.003, ..., 3602.345 which is multiple
seconds jitter after an hour which is not ideal data analysis
as the timestamp for each run is not matched.

This CL fixes that logic to eliminate jitter by calculate sleep
time that does not include the time to log power.

BUG=b:69810407
TEST=No jitter in 1 hour PLT

Change-Id: I64104b317defd78f8c64e5d716b7f1ba8888e1da
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/883777
Reviewed-by: Todd Broch <tbroch@chromium.org>