FAFT: mode_switcher: Adjust to vboot and depthcharge changes
This patch changes FAFT code to be in sync with the recent vboot menu
change that removes the RECOVERY menu from the dev mode transition. It
also removes a superfluous button press that was necessary to counteract
a bug with handling button combos in depthcharge that has now been
fixed.

CQ-DEPEND=CL:884839
BRANCH=None
BUG=None
TEST=None

Change-Id: Ia9d6084f9ddd5ff8b471197593cbabda19de028c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/884840
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Shelley Chen <shchen@chromium.org>