autotest: rpc_interface: make delete_shard simpler
The current delete_shard behavior contains a lot of complicated, slow,
and not very reliable code to try to babysit the state of DUTs during
the shard deletion transition. This code is slow and prone to
breakage, and the value it provides is dubious because DUTs will end up
with eventually correctness anyway.

Delete all the complicated logic, and just make delete_shard delete the
shard.

BUG=chromium:806131, chromium:806129
TEST=None

Change-Id: If8d877bce01c0dbd824746f62160331b7a886cdd
Reviewed-on: https://chromium-review.googlesource.com/888086
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>