chrome_cr50: add function for testing servo v4 dts mode
Any test running with servo v4 may have the type c cable plugged into
the CCD port. This can be used to enable/disable CCD on Cr50 even if CCD
is not being used to run the test. This changes chrome_cr50 to adapt all
of the ccd functions to run with any servo v4 setup. It also adds a
function to check dts mode support.

BUG=none
BRANCH=none
TEST=run firmware_Cr50CCDServoCap

Change-Id: I800176e44a64fa7f4e17359414420c3377e114b2
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/893602
Reviewed-by: Wai-Hong Tam <waihong@google.com>