network_WlanDriver: make it fatal if no devices are found
We want to use this as a litmus test for basic driver support in the BVT
suite. But currently, it gives a Pass (TestNA) for a device with no
Wifi driver loaded. This is bad, since it doesn't catch one of various
failure modes:
(a) driver is not enabled, when it should be
(b) driver is poorly written and fails to initialize reliably
(c) other drivers (e.g., cfg80211.ko duplication) prevent our driver
    from loading

These are things we should catch.

Let's go with the following rule: if you disable shill's Wifi support
(USE="-wifi"), then we'll forgive you. But if you enable Wifi support,
we expect to find a supported Wifi device.

BUG=chromium:793526
TEST=$subject test, on devices with
   (a) unsupported driver -> FAIL
   (b) supported driver, but removed (rmmod) -> FAIL
   (c) no shill/wifi support (USE="-wifi") -> TEST_NA
   (d) supported driver, properly initialized -> PASS

Change-Id: I54277d8ab12f2c65f7a2c7c9aa8cc79956d61e83
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/890121
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>