network_WlanDriver: Add 4.4 support for QCA6174A
This passed AVL on on some 4.4 devices already. We're using 'ar10k' for
now, but we can revisit that (e.g., what comes on 4.14?) later.

Drop the ath10k_core, because that's not the endpoint driver; it's the
core/helper module.

Also, note that I left the commas on the end; it's much nicer to be able
to append to the dictionary without modifying existing lines.

BUG=none
TEST=test_that, on scarlet

Change-Id: I897f1c0a3a65a4e5c6ac37da66b17d5d39e345e5
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/890122
Reviewed-by: Edward Hill <ecgh@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>