Rewrite platform_CryptohomeTPMReOwn
This test is very old and does not make use of any of the
cryptohome-interfacing libraries, which would greatly simplify
the code. In addition a lot of code is duplicated between the
subtests. On top of that this is the last user of the non-async
mount on the CLI so moving to the cryptohome libraries allows us
to remove that method from cryptohome.

The test should also be a little faster since we only need to
reboot the DUT 3 times instead of 7.

BUG=chromium:765525
TEST=run platform_CryptohomeTPMReOwnServer

Change-Id: If472aeed2070bb22b4bcb759a2c3b5025ef637c8
Reviewed-on: https://chromium-review.googlesource.com/895143
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>