Add new firmware_Bmpblk test to FAFT
This patch adds a new FAFT lv4 test to validate whether the firmware
screen images have been correctly scaled for the device's panel
resolution. In the past, we have repeatedly forgot to set this correctly
in the bmpblk board.yaml, which means the respective Chromebooks shipped
with images sized for 1366x768 that get scaled up at runtime and thus
look very blurry. This test should finally prevent that from ever
happening again.

The test itself merely checks for the existence of a flag file in the
firmware image's CBFS. The "heavy lifting" is done by CL:898219 which
makes the chromeos-bmpblk.ebuild insert this flag file into images
whenever it runs with the fallback resolution.

BUG=None
TEST=Ran on Kevin both with and without explicitly scaled images.

Change-Id: Iedc34e29affacbea8daa2523306b252c3575d81b
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/898221
Reviewed-by: Wai-Hong Tam <waihong@google.com>