Use cryptohome libraries in platform_CryptohomeChangePassword
This test rolls its own cryptohome CLI calls for everything. Most
of these are strictly worse versions of the tools in the existing
cryptohome libraries but the call to MigrateKey isn't present
there. So, add that to the cryptohome libraries, and then we can
make this test a lot shorter.

This is the last remaining direct call to --action=mount in the
cryptohome CLI so we can now move the mount_vault function to use
--action=mount_ex and deprecate the legacy mount methods.

BUG=chromium:765525
TEST=run platform_CryptohomeChangePassword

Change-Id: Ie8010223eac4445698b7f4ea0209ce1affe16c8c
Reviewed-on: https://chromium-review.googlesource.com/898545
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>