autoupdate test over cellular connection.
While connected to cellular, the DUT will not be able to access a
devserver. So we need to copy the payload to a publicly accessible
google storage location that we will then download over cellular.

We will use an omaha instance on the DUT (via
autoupdate_CannedOmahaUpdate) that will point to that google storage
location.

We will then check that the update engine logs contain the cellular
entries if the update completes successfully.

Other small things in this CL:
1. I have reworked the get_delta_payload() function to be able to return
either the full or delta payload url.

2. I've add a function which calls the devservers /api/fileinfo API to
return information about staged files (SHA256 and size) because
gs://chromeos-image-archive/ payloads don't have the .json file with this
info.

BUG=chromium:797047
TEST=autoupdate_Cellular.full passes
TEST=autoupdate_Cellular.delta passes

Change-Id: I14cd15615a5606a3ed2ead4721c253baf5212b21
Reviewed-on: https://chromium-review.googlesource.com/900362
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>