video_JpegDecodeAcceleratorPerf: Fix power utility issues
There are 2 issues for power utility.
1. ectool returning non-zero status.
2. There is no power sysfs interface for battery.
This CL fixs both of them.
It will retry ectool command 3 times for non-zero status.
It will check if the battery is valid before use the power status
utility.

BUG=chromium:808332
TEST=Test it on peach_pi for case 1.
Test it on eve for case 2.
Test it on samus for normal case.

Change-Id: I5d0f467c2d317c30d6fcd36b7d27d7693754f0f4
Reviewed-on: https://chromium-review.googlesource.com/901044
Commit-Ready: Hsu Wei-Cheng <mojahsu@chromium.org>
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Hsu Wei-Cheng <mojahsu@chromium.org>