Tweak platform_CryptohomeKeyEviction to avoid proxy
The proxy calls into the D-Bus API directly instead of using the
CLI wrapper. Using the CLI means that all calls go through the
same pathway and we can make changes to the D-Bus API without
having to modify these tests.

In addition, it's difficult to move these calls to MountEx because
protobufs are unwieldy to use here. Also, the CLI utilities will
now unmount the vault before trying to recreate it because the
removal might fail if the vault is still mounted.

BUG=None
TEST=run test

Change-Id: Ib3d6ad11be94776f709703f10a197fe171a96709
Reviewed-on: https://chromium-review.googlesource.com/902687
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>