Rewrite platform_CryptohomeMigrateKey to avoid proxy
Simplify this test by using CLI utilities. Upgrade the
change_password function to raise a ChromiumOSError if the
key migration fails, so we can test whether or not it succeeds
without calling test_auth (which would fail for other reasons
in e.g. the nonexistent user sub-case of this test.)

BUG=None
TEST=run test

Change-Id: Ie60fde85169b7cb3236d4a27d0d48f93aed90fad
Reviewed-on: https://chromium-review.googlesource.com/902798
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>