Forced OOBE Update over cellular.
We already have a forced OOBE update test but we want to have a test for
cellular as a few devices will be shipping with LTE.

Working with cellular means the test can no longer use a devserver in the lab.
It cannot verify the update worked successfully by checking the hostlog.

To make this work it requires enabling updates over cellular in the
server test. Then making the payload public so it can be reached over
cellular. Then checking the update engine log for cellular entries and update
successful entries.

The client test will need to switch to cellular, start an omaha instance
and watch the update until just before reboot.

BUG=chromium:809719
BUG=chromium:809681
TEST=autoupdate_ForcedOOBEUpdate.cellular.{full|delta} passes
TEST=autoupdate_Cellular still passes
TEST=autoupdate_ForcedOOBEUpdate still passes
Change-Id: Iad1b1aadd6f437aebc8026bb3ce612466fa78fb4
Reviewed-on: https://chromium-review.googlesource.com/911605
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>