packager: Delete old file before copying new tarball
Speculative change. In the buildbots, the autotest/packages
directory seems to have stale versions of the files.

This implies that either the compare_checksum() is failing,
or the older files present in autotest/packages aren't getting
overwritten.

So, delete the old version of the file if it exists before
copying the new one.

BUG=chromium:782034
TEST= Ran a local trybot with a dummy autotest CL : http://shortn/_G6zTNPaItE

Change-Id: I4913092cd95000f66a0b14a9cc08347e36e465f2
Reviewed-on: https://chromium-review.googlesource.com/909332
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>