Tweak platform_CryptohomeNonDirs to avoid proxy
This moves another user off CryptohomeProxy so we can get rid of
it and remove AsyncMount. In addition there is a failure message
that is the inverse of what the actual failure issue is, so this
was changed to avoid misleading people reading the test logs.

BUG=chromium:765525
TEST=run test on cyan

Change-Id: I36f4a4225b9c878337af327039ef0c4a57d60691
Reviewed-on: https://chromium-review.googlesource.com/912179
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>