faft: If changing DISABLE_EC_SOFTWARE_SYNC GBB flag, reboot afterwards
At the test startup, it may change the GBB flags to proper ones.
If DISABLE_EC_SOFTWARE_SYNC flag is changed, should reboot the
device, to perform a software sync on the next boot, in order
to get a clear state.

BUG=b:73249665
TEST=Set the GBB flags to enable DISABLE_EC_SOFTWARE_SYNC and
ran the firmware_ECUpdateId test which forced a reboot.

Change-Id: Iec0821d1023a77182fcb4ef1029993925139d661
Reviewed-on: https://chromium-review.googlesource.com/915044
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>