Chameleon: Split extension into two for audio and display APIs
Chrome display APIs require the autotestPrivate permission which can
only be applied to an extension. But the audio APIs can only be used
in an app, not an extension. We need to split the current single
extension/app into two: one as an extension for display APIs and
one as an app for audio APIs.

The display test extension inherits the original multimedia test
extension (whitelisted for autotestPrivate) and the audio test
extension is new.

BUG=chromium:810519
TEST=Ran suite:chameleon_hdmi_perbuild and suite:chameleon_audio_perbuild

Change-Id: I7f9660ad25121f208aed7fa7b40882f90076382f
Reviewed-on: https://chromium-review.googlesource.com/917008
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>