network_WlanDriver: ask the system directly for interfaces, not shill
It's possible for shill to ignore certain network interfaces (e.g., with
moblab, we find that moblab-network-bridge-init is blacklisting wlan0,
among others), causing this test to fail, when in fact, the Wifi driver
is working just fine.

I couldn't find a way to query shill to determine whether a device is
"blacklisted" -- the best I could come up with was to look for a Device
type of "unknown", except this required figuring out the device name
without shill's help (shill doesn't enumerate blacklisted devices for us).

Instead, I figured that because we're really looking for just kernel
support in this test (and not really testing for shill behaviors), I
think it makes sense to just query the kernel directly.

So, let's directly list out /sys/class/net/ and look for Wifi devices.

BUG=chromium:812064
TEST=network_WlanDriver, on guado_moblab, kevin, etc.

Change-Id: I726d74f67e41c19f2bcb4d87737fce7cdabf5fc2
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/919701
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>