cr50_stress_experimental: try ccd open when AP is off
Try to open ccd while the AP is off. Verify it can be done after cr50
sees the AP turn off and after cr50 resets while the AP is off. CCD will
be used while the AP is off. We need to be able to open ccd even if the
AP can't turn on. This test verifies that is possible.

BUG=b:73292631
BRANCH=none
TEST=none

Change-Id: I679e85cf87230ab70aa794169d7284b90b941d26
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/922909
Reviewed-by: Wai-Hong Tam <waihong@google.com>