cros_host: utils: Properly check mosys command exit status
We assumed before that mosys will print nothing to stdout for
unsupported command. But actually there is an exception to that.
On a Chromebox without EC, "mosys ec info -s fw_version" will
print mosys usage guide to stdout.

This CL properly checks command exit status and ignore stdout
when exit status is not zero.

BUG=b:69810407
TEST=manual, Tested on Tidus

Change-Id: I6af3055515c6e533bf771ef7e595f954e29cd8cd
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/927801
Reviewed-by: Todd Broch <tbroch@chromium.org>