Rename test suite to see if that makes it run.
This test suite has not run since being checked in last week.
There are control files that include it.
The suite is included in attribute_whitelist.txt

So the only thing that I can think of is the filename does not match the
NAME variable within the file.

TEST=None
BUG=None

Change-Id: I3da1ffb2a4de420e7e99598dc12995b4c68f165e
Reviewed-on: https://chromium-review.googlesource.com/929802
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>