[autotest] Set higher priority for some CTS tests.
This is just making use of a moblab only feature to allow an individual
test to override the suite priority.

This helps when a long running test initially fails ( provisioning failure
etc ) it gets auto retried, but as the last test in the suite, this will
ensure the test will retry as soon as a DUT is available.

TEST=build stage and run the tests on moblab.
BUG=None

Change-Id: Ied07016c2d66cea5d54c1ff768931447f8c48986
Reviewed-on: https://chromium-review.googlesource.com/932366
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Matt Mallett <mattmallett@chromium.org>