power_SuspendStress: improve, enable network connection checking
We know that this test is extremely flaky, mostly due to network
connectivity. Autotest typically expects to be able to keep an SSH
connection alive while a client test is running. Even if the test is
running completely fine otherwise, it's possible for the SSH connection
to drop, and for the "runner" side to exit and (in _diagnose_dut())
notice it can't ping the device, reporting:

  FAIL: Autotest client terminated unexpectedly: DUT is no longer
  pingable, it may have rebooted or hung.

While in fact, the device is continuing to run the suspend/resume loop,
blissfully unaware that the test runner has aborted.

We can fix (read: paper over) this flakiness a bit by:

(a) increasing the minimum time between resume/suspend attempts
(b) enabling the existing network connection validation within the test
    ('check_connection' was always False before, making this code dead)
(c) adding a timeout loop so that if the connection doesn't come at
    first (e.g., after 5 seconds) we continue retrying for the next 10
    seconds

Altogether, this gives a higher probability of keeping the SSH
connection alive. But it's still not guaranteed.

BUG=b:73349131, b:72341611
TEST=`test_that --autotest_dir ~/trunk/src/third_party/autotest/files/ -b $BOARD $IP f:.*power_SuspendStress/control.bareFSI`
     (with a stable ethernet connection; also with removing the network
     cable partway through)

Change-Id: I72f2a34d6c8c6faad2a001e559b78e17e9e4a4d8
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/932967
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Gwendal Grignou <gwendal@google.com>
Reviewed-by: Todd Broch <tbroch@chromium.org>