Catch cellular TestError failures and surface as TestFail.
When any of the autoupdate tests fail because the sim card was not setup
properly (or is missing etc), they fail with TestError. I want them to
be surfaced as a TestFail in the client test so that the server test
failure message includes the failure reason.

This prevents having to dig into logs to see what happened.

BUG=None
TEST=autoupdate_cellular and autoupdate_forcedoobeupdate.cellular.delta
on DUTs without sim card installed.

Change-Id: I4994916477d3c8ac918bdb1a4236c59f0c9ae674
Reviewed-on: https://chromium-review.googlesource.com/933367
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>