tradefed_test: Do not invalidate the download cache in case of timeout.
We had 538 timeouts in the last 28 days. That is, ~20 times a day we were
clearing out the cache on some shard. This is spoiling the value of the
caching a lot, and possibly making the test flakier by increasing the IO load.

The timeouts are almost always caused by either test failures or device
disconnection, not by the corrupted download cache.

BUG=None
TEST=Run CtsViewTestCases with timeout=60 sec and confirmed it didn't clean up.

Change-Id: Ie1cf73c0948fa36514516a54bda474c3dcf490d8
Reviewed-on: https://chromium-review.googlesource.com/934006
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>