firmware_CorruptRecoveryCache: Ensure init/restore DUT state
This test needs memory training to be done before the start of the
test, but never makes sure that this is the case.  So, it's possible
if the DUT was flashed right before running this test that we fail the
3-key recovery test because we are doing memory training.  Added an
extra recovery boot at the beginning of the test to make sure memory
training is done.  Also, the test never cleans up after it runs and
exits with the DUT in recovery mode because of the 4-key recovery
test.  Added an extra reboot to ensure that DUT is not in recovery
mode.  Additionally, this test was saving and restoring the firmware,
but the test itself never corrupts the firmware.  So, there's no
reason for this to be done.

BUG=b:73293232
BRANCH=None
TEST=run firmware_CorruptRecoveryCache

Change-Id: I5ea34e695f5ab1bb2f2f513da5af851dc701fa46
Signed-off-by: Shelley Chen <shchen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/934146
Reviewed-by: Furquan Shaikh <furquan@chromium.org>