[autotest] Fix lab inventory for "detect repair loops".
When not invoked for debug, the lab inventory script requires at
least one option that specifies one of the available inventory
reports.  The new repair loop detection is an inventory report, but
the check for "at least one report" didn't recognize it as such.
The result was that inventory runs requesting "detect repair loops
only" would fail.

This updates the command line parsing to do the right thing, and
updates the unit tests to test against that case.

BUG=chromium:775199
TEST=run unit tests, run the cron job script manually

Change-Id: I5eb664a2901e3c351fd2efce0d5dd1dfe48175c2
Reviewed-on: https://chromium-review.googlesource.com/937782
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Paul Hobbs <phobbs@google.com>