power_suspend: clear powerd's HWCLOCK file in between iterations
After the first suspend iteration, powerd will have a leftover timestamp
in /run/power_manager/root/hwclock-on-resume. But on subsequent
iterations, powerd doesn't guarantee that this file gets updated, e.g.,
if something (like `hwclock` itself) fails. This can confuse things on
subsequent runs and cause reported "spurious wakeups" (because we
measure time against the old timestamp).

Just clear the file before we suspend, to avoid any issues with
staleness. The existing parsing code will handle this properly (abort,
with a "Broken RTC timestamp") if it remains empty.

We could probably also convince powerd to clear this file in error
conditions, but this makes for a fine backup option.

BUG=chromium:815249
TEST=power_SuspendStress with misbehaving RTC

Change-Id: I53af2375614b352ce8006b78d8d3d562af612027
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/940601
Reviewed-by: Dan Erat <derat@chromium.org>