faft: Swap servo v4 role to SNK for recovery USB boot
In recovery boot, the locked EC RO doesn't support PD for most
of the CrOS devices. The default servo v4 power role is a SRC.
The DUT becomes a SNK. Lack of PD makes CrOS unable to do the
data role swap from UFP to DFP; as a result, DUT can't see the
USB disk and the Ethernet dongle on servo v4.

This is a workaround to set servo v4 as a SNK, for every FAFT
test which boots into the USB disk in the recovery mode.

A test should define the argument used_for_recovery=True in
the setup_usbkey().

BUG=b:123719781
TEST=Only use a single-USB-port on a locked Snappy device,
which connects to servo v4 with servo micro, and servo v4
connects to Ethernet, USB disk, and power supply.
 * Run the FAFT test firmware_RecoveryButton.ec_wp which locks
   EC through WP and boots into the servo v4 USB disk.
 * Run the firmware_PDProtocol not broken.

Change-Id: Id92f6fa8a87405880c17f2853b40dfb2568fce63
Reviewed-on: https://chromium-review.googlesource.com/943832
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>