network_WlanDriver: add new .bvt version, with board exceptions
Some boards have Wifi which is unresolvably flaky, such that the driver
may just crash entirely and leave us with no Wifi device. We don't want
this flakiness to affect CQ results.

For now, we run this test only on bvt-perbuild (as we were doing
already), but we plan to move to bvt-cq soon.

The current exception list only contains nyan_kitty, because it's the
only lab device currently failing with unresolvable symptoms in the lab.
Occasionally, the device fails to load the mwifiex firmware correctly,
saying "FW failed to be active in time" after some long timeout. We've
debugged this issue before on some later platforms, but this requries
extensive vendor assistance, as the bug lies entirely within the
firmware.

I wouldn't be surprised if we had to add other Marvell-Wifi devices in
the future.

Also adjust the DOC wording to more accurately describe what it is
testing.

BUG=chromium:793526
TEST=`test_that ... network_WlanDriver.bvt` with $BOARD in and out of
     exception list; with and without driver loaded; see WARN result for
     forgiven failures, GOOD for success

Change-Id: I8107f0d3b9a9d520009407a60dbdb43cf869b0e1
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/944502
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>