memory_qual: Add new control file for moblab
moblab can't run 10K suspend loops as test will fail because
moblab can't SSH in to the DUT.

Before this we use moblab control file editing feature to manually
change loop count to 2 but this feature is dropped.

This CL fixes that by add new control file to run this test in moblab.

BUG=b:74078789
TEST=None

Change-Id: I29fef7d8338aa8cf69f4b0f9e91eeda409cc5b78
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/944254
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>