Test that we continue with second URL when first fails.
This is an automated test to replace the manual execution of this test:
https://testtracker.googleplex.com/testplans/testcase/detail/906000?id=450

I've addd a server test which gets the current payload to use for the
build the test is running on. This class will be used by all tests that
are testing something in the omaha response (e.g Two UrlBase lines,
bad MetaDatasize, bad SHA256 etc)

The client test does all of the functionality testing.

I have also updated our nano omaha devserver to switch to the other url
after one failure.

BUG=chromium:825958
TEST=autoupdate_OmahaResponse.url_switch.full
TEST=autoupdate_OmahaResponse.local

Change-Id: I37485dd54eb5c4d714309056b281c0cd702fa595
Reviewed-on: https://chromium-review.googlesource.com/985136
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>