[autotest] Pass the upload logs dir via `arguments`.
In the DUT deployment tools, determining the directory where logs
should be uploaded was done by calling a public function in the
`commandline` module, passing the parsed command arguments.

This changes the flow to store the directory name in the parsed
arguments when they're validated, eliminating the public function
in favor of a public field.

BUG=chromium:632751
TEST=`deployment_test -nst`

Change-Id: I763b81b4c89a84351d4050e08635f7c34cd6f21d
Reviewed-on: https://chromium-review.googlesource.com/952663
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>